rangiffler-api-backend-dev-2 Осуществлена реорганизация аннотаций и тестов, улучшена работа с данными пользователей и тестируемость интерфейсов:
- Переработана аннотация `@User` и заменена на `@UserType` для унификации.
- Удален `UsersQueueExtension`, его функционал перенесен в `UserExtension` с уточненным генерированием тестовых данных и автоматической очисткой.
- Расширено использование новых аннотаций и расширений в тестах (`RegistrationTests`, `LoginTests`, `FriendsWebTests`) для работы с типизированными пользователями.
- Перемещены модели и сущности данных в более структурированные пакеты (например, `io.student.rangiffler.model` и `io.student.rangiffler.data.entity.*`).
- Добавлена утилита `RandomUtils` для генерации случайных строк в тестах.
- Внедрен и настроен плагин `dgs-codegen` в проекте для генерации типизированных классов по GraphQL-схемам.
- Поддержка автоматического маппинга типов и улучшений в сборке Java в `build.gradle.kts`.
