+++ rangiffler-e2e/src/test/java/io/student/rangiffler/data/entity/userdata/UdUserEntity.java
@@ -1,4 +1,4 @@
-package io.student.rangiffler.data.userdata;
+package io.student.rangiffler.data.entity.userdata;
 
 import lombok.Data;
