--- rangiffler-e2e/src/test/java/io/student/rangiffler/jupiter/annotation/User.java
@@ -1,15 +1,0 @@
-package io.student.rangiffler.jupiter.annotation;
-
-import io.student.rangiffler.jupiter.extension.UserExtension;
-import org.junit.jupiter.api.extension.ExtendWith;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-@ExtendWith(UserExtension.class)
-public @interface User {
-}
