+++ rangiffler-e2e/src/test/java/io/student/rangiffler/jupiter/annotation/UserType.java
@@ -1,5 +1,8 @@
 package io.student.rangiffler.jupiter.annotation;
 
+import io.student.rangiffler.jupiter.extension.UserExtension;
+import org.junit.jupiter.api.extension.ExtendWith;
+
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -7,6 +10,7 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.PARAMETER})
 @Retention(RetentionPolicy.RUNTIME)
+@ExtendWith(UserExtension.class)
 public @interface UserType {
     Type value() default Type.EMPTY;
