+++ rangiffler-e2e/src/test/java/io/student/rangiffler/service/UserDbClient.java
@@ -3,10 +3,10 @@ package io.student.rangiffler.service;
 import com.atomikos.jdbc.AtomikosDataSourceBean;
 import com.mysql.cj.jdbc.MysqlXADataSource;
 import io.student.rangiffler.config.Config;
-import io.student.rangiffler.data.auth.AuthUserEntity;
-import io.student.rangiffler.data.auth.AuthorityEntity;
-import io.student.rangiffler.data.userdata.UdUserEntity;
-import io.student.rangiffler.models.UserJson;
+import io.student.rangiffler.data.entity.auth.AuthUserEntity;
+import io.student.rangiffler.data.entity.auth.AuthorityEntity;
+import io.student.rangiffler.data.entity.userdata.UdUserEntity;
+import io.student.rangiffler.model.UserJson;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.ResultSetExtractor;
 import org.springframework.jdbc.datasource.DriverManagerDataSource;
