+++ rangiffler-e2e/src/test/java/io/student/rangiffler/service/UsersClient.java
@@ -1,6 +1,6 @@
 package io.student.rangiffler.service;
 
-import io.student.rangiffler.models.UserJson;
+import io.student.rangiffler.model.UserJson;
 
 public interface UsersClient {
     UserJson createUser(String username, String password);
