+++ rangiffler-e2e/src/test/java/io/student/rangiffler/tests/FriendsWebTests.java
@@ -3,20 +3,20 @@ package io.student.rangiffler.tests;
 import com.codeborne.selenide.Selenide;
 import io.student.rangiffler.config.Config;
 import io.student.rangiffler.jupiter.annotation.UserType;
-import io.student.rangiffler.jupiter.extension.UsersQueueExtension;
+import io.student.rangiffler.jupiter.extension.UserExtension;
 import io.student.rangiffler.page.LoginPage;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import static io.student.rangiffler.jupiter.annotation.UserType.Type.*;
 
-@ExtendWith(UsersQueueExtension.class)
+@ExtendWith(UserExtension.class)
 public class FriendsWebTests extends BaseTest {
 
     private static final Config CFG = Config.getInstance();
 
     @Test
-    public void friendShouldBePresentInFriendsTable(@UserType(WITH_FRIEND) UsersQueueExtension.StaticUser user) {
+    public void friendShouldBePresentInFriendsTable(@UserType(WITH_FRIEND) UserExtension.TestUser user) {
         Selenide.open(CFG.frontUrl(), LoginPage.class)
                 .clickLoginBtn()
                 .login(user.username(), user.password())
@@ -29,7 +29,7 @@ public class FriendsWebTests extends BaseTest {
 
     @Test
 //    @DisabledByIssue("3")
-    public void friendsTableShouldBeEmptyForNewUser(@UserType(EMPTY) UsersQueueExtension.StaticUser user) {
+    public void friendsTableShouldBeEmptyForNewUser(@UserType(EMPTY) UserExtension.TestUser user) {
         Selenide.open(CFG.frontUrl(), LoginPage.class)
                 .clickLoginBtn()
                 .login(user.username(), user.password())
@@ -39,7 +39,7 @@ public class FriendsWebTests extends BaseTest {
     }
 
     @Test
-    public void incomeInvitationBePresentInFriendsTable(@UserType(WITH_INCOME_REQUEST) UsersQueueExtension.StaticUser user) {
+    public void incomeInvitationBePresentInFriendsTable(@UserType(WITH_INCOME_REQUEST) UserExtension.TestUser user) {
         Selenide.open(CFG.frontUrl(), LoginPage.class)
                 .clickLoginBtn()
                 .login(user.username(), user.password())
@@ -51,7 +51,7 @@ public class FriendsWebTests extends BaseTest {
     }
 
     @Test
-    public void outcomeInvitationBePresentInAllPeoplesTable(@UserType(WITH_OUTCOME_REQUEST) UsersQueueExtension.StaticUser user) {
+    public void outcomeInvitationBePresentInAllPeoplesTable(@UserType(WITH_OUTCOME_REQUEST) UserExtension.TestUser user) {
         Selenide.open(CFG.frontUrl(), LoginPage.class)
                 .clickLoginBtn()
                 .login(user.username(), user.password())
