+++ rangiffler-e2e/src/test/java/io/student/rangiffler/tests/LoginTests.java
@@ -2,8 +2,8 @@ package io.student.rangiffler.tests;
 
 import com.codeborne.selenide.Selenide;
 import io.student.rangiffler.config.Config;
-import io.student.rangiffler.jupiter.annotation.User;
-import io.student.rangiffler.models.UserJson;
+import io.student.rangiffler.jupiter.annotation.UserType;
+import io.student.rangiffler.jupiter.extension.UserExtension;
 import io.student.rangiffler.page.LoginPage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -12,9 +12,8 @@ public class LoginTests extends BaseTest {
     private static final Config CFG = Config.getInstance();
 
     @Test
-    @User
     @DisplayName("Успешная авторизация")
-    void mainPageShouldBeDisplayedAfterSuccessfulLogin(UserJson user) {
+    void mainPageShouldBeDisplayedAfterSuccessfulLogin(@UserType() UserExtension.TestUser user) {
         Selenide.open(CFG.frontUrl(), LoginPage.class)
                 .clickLoginBtn()
                 .login(user.username(), user.password())
@@ -22,9 +21,8 @@ public class LoginTests extends BaseTest {
     }
 
     @Test
-    @User
     @DisplayName("|-| Неуспешная авторизация с неправильным паролем")
-    public void userShouldStayOnLoginPageAfterLoginWithBadCredentials(UserJson user) {
+    public void userShouldStayOnLoginPageAfterLoginWithBadCredentials(@UserType() UserExtension.TestUser user) {
         Selenide.open(CFG.frontUrl(), LoginPage.class)
                 .clickLoginBtn()
                 .setUsername(user.username())
